% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arity-package.R
\docType{package}
\name{arity-package}
\alias{arity}
\alias{arity-package}
\title{arity: Format R Code with Arity}
\description{
R bindings to the deterministic, rule-based formatter from arity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/arity-r}
  \item \url{https://arity.cc}
  \item Report bugs at \url{https://github.com/jolars/arity-r/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johan@jolars.co} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Johan Larsson \email{johan@jolars.co} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID}) [copyright holder]
}

}
\keyword{internal}
