% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{format_config_native}
\alias{format_config_native}
\title{Resolve formatter settings from an explicit or discovered configuration.}
\usage{
format_config_native(directory, config, no_config, fallback)
}
\description{
Resolve formatter settings from an explicit or discovered configuration.
}
\keyword{internal}
