% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{format_text_native}
\alias{format_text_native}
\title{Format one R source string.}
\usage{
format_text_native(
  source,
  line_width,
  indent_width,
  line_ending,
  roxygen_markdown,
  verify
)
}
\description{
Format one R source string.
}
\keyword{internal}
